+++ homura_art/__main__.py
@@ -1,26 +1,22 @@
-import os
 import random
 import shutil
 import subprocess
 import sys
 
-from peewee import fn
 from PySide6.QtWidgets import (
-    QApplication,
     QMainWindow,
     QDialog,
-    QSizePolicy
+    QApplication,
 )
-from PySide6.QtGui import QPixmap, QResizeEvent
+from PySide6.QtGui import QPixmap, QResizeEvent, QImageReader
 from PySide6.QtCore import Qt
 
-from homura_art.model import File
+from homura_art.model import File, FilePost, Post
 from homura_art.ui.main_window import Ui_MainWindow
 from homura_art.ui.collage_preview import Ui_CollagePreview
-from homura_art.utilities import sync
-import hydrus_api
+from homura_art.utilities import get_hash, sync
+import shutil
 
-ACCCESS_KEY = os.environ.get("ACCCESS_KEY", "")
 
 class CollagePreview(QDialog, Ui_CollagePreview):
     def __init__(self) -> None:
@@ -39,72 +35,148 @@ class CollagePreview(QDialog, Ui_CollagePreview):
     def resizeEvent(self, event: QResizeEvent) -> None:
         self.repaint()
         return super().resizeEvent(event)
-        
 
 
 class MainWindow(QMainWindow, Ui_MainWindow):
+    def generate_queue(self):
+        queue = []
+        posts = list(Post.select().where(Post.filtered == False))
+        if len(posts) < 24:
+            sync()
+        # files
+        files = list(File.select())
+        random.shuffle(files)
+        if len(posts) < 24:
+            need = min(32 - len(posts), len(files))
+        else:
+            need = 8
+        while files:
+            file = files.pop()
+            queue.append({"path": file.path, "type": "file", "file": file})
+            print(len(queue), need)
+            if len(queue) == need:
+                break
+        # post
+        random.shuffle(posts)
+        while posts:
+            post = posts.pop()
+            path = post.path
+            if not path:
+                continue
+            queue.append({"path": post.path, "type": "post", "post": post})
+            if len(queue) == 32:
+                break
+        random.shuffle(queue)
+        return queue
+
     def __init__(self):
         super(MainWindow, self).__init__()
         self.setupUi(self)
-        self.client = hydrus_api.Client(ACCCESS_KEY)
-        self.files, self.pathes = self.get_random_files()
-        self.left_image = QPixmap(self.pathes[0])
-        self.right_image = QPixmap(self.pathes[1])
+        self.new_queue = []
+        self.queue = self.generate_queue()
+        self.left = self.queue.pop()
+        self.right = self.queue.pop()
+        self.left_image = QPixmap(self.left["path"])
+        self.right_image = QPixmap(self.right["path"])
+
         self.left_win_button.clicked.connect(self.left_win)
         self.right_win_button.clicked.connect(self.right_win)
         self.tie_button.clicked.connect(self.tie)
-        self.skip_button.clicked.connect(self.get_next)
         self.delete_left_button.clicked.connect(self.delete_left)
         self.delete_right_button.clicked.connect(self.delete_right)
         self.collage_button.clicked.connect(self.get_collage)
 
-
-    def delete(self, file):
-        self.client.delete_files([file.hash_str])
-        file.delete_instance()
-        self.get_next()
+    def delete(self, item):
+        if item["type"] == "post":
+            item["post"].filtered = True
+            item["post"].save()
+        if item["type"] == "file":
+            item["file"].delete_instance(recursive=True)
+        if not self.queue:
+            if len(self.new_queue) >= 2:
+                self.queue = self.new_queue
+                self.new_queue = []
+            else:
+                self.queue = self.generate_queue()
+            print("queue:", len(self.queue), "new queue:", len(self.new_queue))
+        item = self.queue.pop()
+        print("queue:", len(self.queue), "new queue:", len(self.new_queue))
+        return item
 
     def delete_left(self):
-        self.delete(self.files[0])
+        self.left = self.delete(self.left)
+        self.left_image = QPixmap(self.left["path"])
+        self.repaint()
 
     def delete_right(self):
-        self.delete(self.files[1])
+        self.right = self.delete(self.right)
+        self.right_image = QPixmap(self.right["path"])
+        self.repaint()
 
-    def play(self, result):
+    def archive(self, item):
+        if item["type"] == "file":
+            return item
+        else:
+            hash = get_hash(item["path"])
+            file = File.get_or_none(File.hash == hash)
+            post = item["post"]
+            if file:
+                FilePost.create(file=file, post=post)
+            else:
+                path = item["path"]
+                file = File.create(hash=hash, ext=path.suffix)
+                new_path = file.path
+                new_path.parent.mkdir(parents=True, exist_ok=True)
+                shutil.move(path, new_path, copy_function=shutil.copy2)
+                FilePost.create(file=file, post=post)
+            post.filtered = True
+            post.save()
+            return {"path": file.path, "type": "file", "file": file}
+
+    def play(self, rating_a, rating_b, result):
         def get_expected_score(r_a, r_b):
             return 1 / (1 + 10 ** ((r_a - r_b) / 400))
 
         def get_new_rating(r_a, r_b, score):
             return r_a + 32 * (score - get_expected_score(r_a, r_b))
 
-        self.files[0].rating = get_new_rating(
-            self.files[0].rating, self.files[1].rating, result
+        result_a = get_new_rating(rating_a, rating_b, result)
+        result_b = get_new_rating(rating_b, rating_a, 1 - result)
+        return result_a, result_b
+
+    def win(self, result):
+        self.left = self.archive(self.left)
+        self.right = self.archive(self.right)
+        if result == 1:
+            self.new_queue.append(self.left)
+        elif result == 0:
+            self.new_queue.append(self.right)
+        self.left["file"].rating, self.right["file"].rating = self.play(
+            self.left["file"].rating, self.right["file"].rating, result
         )
-        self.files[1].rating = get_new_rating(
-            self.files[1].rating, self.files[0].rating, 1 - result
-        )
-        self.files[0].save()
-        self.files[1].save()
-        self.client.archive_files([file.hash_str for file in self.files])
-        print(self.files)
+        self.left["file"].save()
+        self.right["file"].save()
+        if len(self.queue) < 2:
+            if len(self.new_queue) > 1:
+                self.queue = self.new_queue + self.queue
+                self.new_queue = []
+            else:
+                self.queue = self.generate_queue()
+        print("queue:", len(self.queue), "new queue:", len(self.new_queue))
+        self.left = self.queue.pop()
+        self.right = self.queue.pop()
+        self.left_image = QPixmap(self.left["path"])
+        self.right_image = QPixmap(self.right["path"])
+        self.repaint()
 
     def left_win(self):
-        self.play(1)
-        self.get_next()
+        self.win(1)
 
     def right_win(self):
-        self.play(0)
-        self.get_next()
+        self.win(0)
 
     def tie(self):
-        self.play(0.5)
-        self.get_next()
-
-    def get_next(self):
-        self.files, self.pathes = self.get_random_files()
-        self.left_image = QPixmap(self.pathes[0])
-        self.right_image = QPixmap(self.pathes[1])
-        self.repaint()
+        self.win(0.5)
 
     def repaint(self):
         w = self.left_file.width()
@@ -120,48 +192,15 @@ class MainWindow(QMainWindow, Ui_MainWindow):
         self.repaint()
         return super().resizeEvent(event)
 
-    def get_random_files(self):
-        allow_extensions = [".png", ".jpg", ".webp"]
-        while True:
-            ok = True
-    
-            files = File.select().order_by(fn.Random()).limit(2)
-            hashes = [file.hash_str for file in files]
-            extensions = [data["ext"] for data in self.client.get_file_metadata(hashes)]
-            responses = [self.client.get_file(hash) for hash in hashes]
-            for extension in extensions:
-                if extension not in allow_extensions:
-                    ok = False
-            pathes = []
-            for response, extension, name in zip(
-                responses, extensions, ["left", "right"]
-            ):
-                path = f"/tmp/{name}{extension}"
-                pathes.append(path)
-                with open(path, "wb") as out_file:
-                    shutil.copyfileobj(response.raw, out_file)
-            if ok:
-                return files, pathes
 
-    
     def get_collage(self):
-        files = list(File.select().where(File.used == False).order_by(File.rating.desc()).limit(10))
+        files = list(File.select().where(File.used == False).order_by(File.rating.desc()).limit(16))
         files = random.sample(files, 3)
-        hashes = list(file.hash_str for file in files)
-        extensions = [data["ext"] for data in self.client.get_file_metadata(hashes)]
-        responses = [self.client.get_file(hash) for hash in hashes]
-        pathes = []
-        for response, extension, name in zip(
-            responses, extensions, ["1", "2", "3"]
-        ):
-            path = f"/tmp/{name}{extension}"
-            pathes.append(path)
-            with open(path, "wb") as out_file:
-                shutil.copyfileobj(response.raw, out_file)
-        cmd = f"montage {' '.join(pathes)} -geometry x1080 -mode Concatenate /tmp/collage.png"
+        paths = [str(file.path) for file in files]
+        cmd = f"montage {' '.join(paths)} -geometry x1080 -mode Concatenate /tmp/collage.png"
 
         subprocess.call(cmd, shell=True)
-        
+
         if CollagePreview().exec():
             for file in files:
                 file.used = True
@@ -171,16 +210,12 @@ class MainWindow(QMainWindow, Ui_MainWindow):
 
 
 def main():
-    pass
-#     if ACCCESS_KEY:
-#         sync(ACCCESS_KEY)
-#         app = QApplication([])
-#         widget = MainWindow()
-#         widget.show()
-#         widget.resize(1025, 768)
-#         sys.exit(app.exec())
-#     else:
-#         print("Export ACCCESS_KEY before run")
+    QImageReader.setAllocationLimit(0)
+    app = QApplication([])
+    widget = MainWindow()
+    widget.show()
+    widget.resize(1025, 768)
+    sys.exit(app.exec())
 
 
 if __name__ == "__main__":