+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## [Unreleased]
 ### Added
-- Basic model.
-- Sync of data with hydrus network.
+- Allow to add file from remote source ([#27](https://github.com/nikohonu/shiki-organizer/issues/#27)).
+- Subscriptions ([#27](https://github.com/nikohonu/shiki-organizer/issues/#27)).
+- hydrus network as source ([#27](https://github.com/nikohonu/shiki-organizer/issues/#27)).
 - Basic gui, that allow rate image by elo.
 - Generate collage button.